requirements.txt
aioredis==1.3.0
asgi-redis==1.4.3
asgiref==3.2.3
async-timeout==3.0.1
attrs==19.3.0
autobahn==19.11.1
Automat==0.8.0
Babel==2.7.0
beautifulsoup4==4.8.1
cffi==1.13.2
channels==2.3.1
channels-redis==2.4.1
constantly==15.1.0
cryptography==2.8
daphne==2.3.0
Django==2.0.4
django-countries==5.5
django-extensions==2.2.5
django-htmlmin==0.11.0
django-js-asset==1.2.2
django-mptt==0.10.0
django-phonenumber-field==3.0.1
djangorestframework==3.8.2
extensions==0.4
hiredis==1.0.1
html5lib==1.0.1
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
msgpack==0.6.2
msgpack-python==0.5.6
phonenumbers==8.10.22
Pillow==6.2.1
psycopg2==2.8.4
pycparser==2.19
pygraphviz==1.5
PyHamcrest==1.9.0
pytz==2019.3
redis==2.10.6
six==1.13.0
soupsieve==1.9.5
Twisted==19.7.0
txaio==18.8.1
webencodings==0.5.1
zope.interface==4.6.0